Bug in parallel grid sim
Error thrown if there were no events detected by detect_PBE, in which case trialResults.events=[]

Solution: skip Bayesian decoding if isempty(trialResults.events)

+++ functions/parallelize_parameter_tests_2_PF.m
@@ -100,7 +100,7 @@ function [avg_mat, allResults, PFresults] = parallelize_parameter_tests_2(parame
         for ithEnv = 1:pfsim.nEnvironments
             
             PFresults{ithNet}{ithEnv}.E_indices = network.E_indices;
-
+            PFresults{ithNet}{ithEnv}.cluster_mat = network.cluster_mat;
                     
             for ithTrial = 1:pfsim.nTrials
                 %rng(ithTrial)
@@ -248,41 +248,46 @@ function [avg_mat, allResults, PFresults] = parallelize_parameter_tests_2(parame
             
             
             %% Calculate and save preplay Bayesian decoding
-            % 1) save SJlab formated files to tmp dir
-            [filepath,name] = fileparts(tempname);
-            animalprefix = name; save_path = [filepath, filesep, animalprefix, '_direct']; 
-            mkdir(save_path)
+            if ~isempty(trialResults.events) % only decode if there are events
+                % 1) save SJlab formated files to tmp dir
+                [filepath,name] = fileparts(tempname);
+                animalprefix = name; save_path = [filepath, filesep, animalprefix, '_direct']; 
+                mkdir(save_path)
+
+                day = 1; eprun = 2; tet = 1;
+                tetinfo{day}{eprun}{tet}.numcells = parameters.n_E;
+                if numel(linfields{day})<eprun
+                    linfields{day}{eprun} = linfields{day}{1};
+                end
+                if numel(linfields{day}{eprun}{:})==parameters.n
+                    linfields{day}{eprun}{tet}(network.I_indices) = [];
+                end
             
-            day = 1; eprun = 2; tet = 1;
-            tetinfo{day}{eprun}{tet}.numcells = parameters.n_E;
-            if numel(linfields{day})<eprun
-                linfields{day}{eprun} = linfields{day}{1};
+                ripple{day}{eprun}.starttime = trialResults.events(:,1).*parameters.dt;
+                ripple{day}{eprun}.endtime = trialResults.events(:,2).*parameters.dt;
+                for cellid = 1:numel(network.E_indices)
+                    % spikes{day}{eprun}{tet}{cellid}.data(:,1) = V_m(network.E_indices(cellid),:)>= parameters.V_th; % spike times
+                    spikes{day}{eprun}{tet}{cellid}.data(:,1) = find(E_spikes_V_m(cellid,:)).*parameters.dt; % spike times
+                end
+                clear E_spikes_V_m
+
+                save([filepath, filesep, animalprefix, '_direct/', animalprefix, 'tetinfo.mat'], 'tetinfo')
+                save([filepath, filesep, animalprefix, '_direct/', animalprefix, 'linfields01.mat'], 'linfields')
+                save([filepath, filesep, animalprefix, '_direct/', animalprefix, 'rippletime01.mat'], 'ripple')
+                save([filepath, filesep, animalprefix, '_direct/', animalprefix, 'spikes01.mat'], 'spikes')
+                clear tetinfo linfields ripple spikes
+
+                % 2) Save results: 
+                day=1; ep=2; cellcountthresh = 5; savedata = 0; figopt = 0; shuffleIterations = 500;
+                warning off
+                decodingResults = preplay_decoding_CA1_singleday(animalprefix,day,ep,cellcountthresh,[filepath, filesep],savedata,figopt,shuffleIterations);
+                warning on
+                allResults{ithNet}{ithTest}.replaytrajectory = decodingResults{day}{ep}; 
+                % 3) Delete tmp dir files:
+                rmdir(save_path, 's')
+            else
+                allResults{ithNet}{ithTest}.replaytrajectory = [];
             end
-            if numel(linfields{day}{eprun}{:})==parameters.n
-                linfields{day}{eprun}{tet}(network.I_indices) = [];
-            end
-            ripple{day}{eprun}.starttime = trialResults.events(:,1).*parameters.dt;
-            ripple{day}{eprun}.endtime = trialResults.events(:,2).*parameters.dt;
-            for cellid = 1:numel(network.E_indices)
-                % spikes{day}{eprun}{tet}{cellid}.data(:,1) = V_m(network.E_indices(cellid),:)>= parameters.V_th; % spike times
-                spikes{day}{eprun}{tet}{cellid}.data(:,1) = find(E_spikes_V_m(cellid,:)).*parameters.dt; % spike times
-            end
-            clear E_spikes_V_m
-            
-            save([filepath, filesep, animalprefix, '_direct/', animalprefix, 'tetinfo.mat'], 'tetinfo')
-            save([filepath, filesep, animalprefix, '_direct/', animalprefix, 'linfields01.mat'], 'linfields')
-            save([filepath, filesep, animalprefix, '_direct/', animalprefix, 'rippletime01.mat'], 'ripple')
-            save([filepath, filesep, animalprefix, '_direct/', animalprefix, 'spikes01.mat'], 'spikes')
-            clear tetinfo linfields ripple spikes
-            
-            % 2) Save results: 
-            day=1; ep=2; cellcountthresh = 5; savedata = 0; figopt = 0; shuffleIterations = 500;
-            warning off
-            decodingResults = preplay_decoding_CA1_singleday(animalprefix,day,ep,cellcountthresh,[filepath, filesep],savedata,figopt,shuffleIterations);
-            warning on
-            allResults{ithNet}{ithTest}.replaytrajectory = decodingResults{day}{ep}; 
-            % 3) Delete tmp dir files:
-            rmdir(save_path, 's')
             
             %% Main output statistics
             %outputVec(1) = allResults{ithNet}{ithTest}.fracFire; % fraction of spiking neurons over entire simulation
